1 - Testes de Unidade/Features.Tests/01 - Traits/ClienteTests.cs
using Features.Clientes;
using System;
using Xunit;

namespace Features.Tests.Traits
{
    public class ClienteTests
    {
        [Fact(DisplayName = "Cliente - Deve Estar Válido")]
        [Trait("Categoria", "Trait - ClienteTests")]
        public void Cliente_NovoCliente_DeveEstarValido()
        {
            // Arrange
            var cliente = new Cliente(
                Guid.NewGuid(),
                "Eduardo",
                "Pires",
                DateTime.Now.AddYears(-30),
                "edu@edu.com",
                true,
                DateTime.Now);

            // Act
            var result = cliente.EhValido();

            // Assert
            Assert.True(result);
            Assert.Empty(cliente.ValidationResult.Errors);
        }

        [Fact(DisplayName = "Cliente - Deve Estar Inválido")]
        [Trait("Categoria", "Trait - ClienteTests")]
        public void Cliente_NovoCliente_DeveEstarInvalido()
        {
            // Arrange
            var cliente = new Cliente(
                Guid.NewGuid(),
                "",
                "",
                DateTime.Now,
                "edu2edu.com",
                true,
                DateTime.Now);

            // Act
            var result = cliente.EhValido();

            // Assert
            Assert.False(result);
            Assert.NotEmpty(cliente.ValidationResult.Errors);
        }
    }
}
